v3/requirements.txt
mailjet_rest
pandas
google-cloud-bigquery
lxml
yfinance
pyarrow
pyyaml